Print --> SetTect
the print function no longer works so to print the position to the brain I get the value of the x and y position and assign those positions to string variables which can be used as arguments in the set_text function to see where the robot is so that we can re-do all the PID tuning since the robot was re-built. Also, I changed the turnTest to DriveTest since we need to make sure the robot can move forwards since last time, it wasn't.